Backend shopps import bug

+++ symfony/apps/backend/modules/shops/actions/actions.class.php
@@ -120,13 +120,10 @@ class shopsActions extends sfActions
 			set_time_limit(10000);
 
 			$stuff = file_get_contents($request->getParameter('import_url'));
-			die('este: ' . $stuff);
-			
-			//simplexml_load_string($data)
-//			if (false == $products = simplexml_load_file($request->getParameter('import_url'))){
-//				die('A aparut o eroare la deschiderea feed-ului!');
-//				$this->redirect($this->generateUrl('default', array('module' => 'shops', 'action' => 'films')) . '?id=' . $this->shop->getId());
-//			}
+			if (false == $products = simplexml_load_string($stuff)){
+				die('A aparut o eroare la deschiderea feed-ului!');
+				$this->redirect($this->generateUrl('default', array('module' => 'shops', 'action' => 'films')) . '?id=' . $this->shop->getId());
+			}
 
 			/* delete the existing films from this shop */
 			Doctrine_Core::getTable('ShopFilm')->deleteByShop($this->shop->getId());
